Add support for UVC RGB24 format, and dithered YUY2 (default)

+++ firmware/src/core_intercom/shared_memory.hh
@@ -21,7 +21,7 @@ struct SharedMemoryS {
 		ConcurrentBuffer *console_a7_0_buff;
 		ConcurrentBuffer *console_a7_1_buff;
 		ConcurrentBuffer *console_m4_buff;
-		uint16_t *uvc_framebuffer;
+		uint8_t *uvc_framebuffer;
 		DeviceSettingsMessage *icc_device_settings_message;
 	};
